Convert http_request_completion_t headers to multimap
Headers are allowed to have duplicate keys for a given request.
A notable example is the set-***REMOVED*** key.  Without support for this,
the map will collapse these to the last entry provided.

This is an interface breaking change for anyone using the headers
field directly.

This also introduces get_headers_list and get_headers_count.
get_headers is unchanged but it is possible it will return a
different header depending on the underlying implementation of
std::multimap when multiple entries are present.

get_headers_count shows the number of entries with a given key.
get_headers_list just collects all the entries with a given key as a list.

Instead of `headers[key] = value`
use `headers.emplace(key, value)`

Instead of accessing `headers[key]` or `headers.at(key)`, use:
```
auto range = headers.equal_range(key);
if ( range.first != headers.end() )
    for ( auto i = range.first; i != range.second; ++i ) {
        list.emplace_back(i->second);
    }
} else {
    // Empty
}
```